+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `@microsoft/eslint-plugin-sdl` eslint plugin
 
+### Fixed
+
+- section validation inside repeatable sets and nested forms
+
 ## [2.11.0] - 2023-04-14
 
 ### Added