+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `isDirty` validation for Checkbox Element
+
 ### Added
 
 - Repeatable set min/max entries based on another element